backend/pyproject.toml
[tool.poetry]
# replace name 'template' with the name of the project
name = "template"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
SQLAlchemy = "^1.3.20"
python-dotenv = "^0.15.0"
Flask-SQLAlchemy = "^2.4.4"
psycopg2-binary = "^2.8.6"
alembic = "^1.4.3"
Flask-Cors = "^3.0.9"
Flask-RESTful = "^0.3.8"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
